Rework point cloud 3D editing tools
Qgs3DMapToolPointCloudChangeAttribute becomes new base class for polygon
tool and paintbrush tool. Processing method has been moved to new
Qgs3DEditUtils class.
