+++ src/app/3d/qgs3dmaptoolpolygon.h
@@ -1,9 +1,9 @@
 /***************************************************************************
-  qgs3dmaptoolpaintbrush.h
-  --------------------------------------
-  Date                 : January 2025
-  Copyright            : (C) 2025 by Matej Bagar
-  Email                : matej dot bagar at lutraconsulting dot co dot uk
+    qgs3dmaptoolpolygon.h
+    ---------------------
+    begin                : February 2025
+    copyright            : (C) 2025 by Matej Bagar
+    email                : matej dot bagar at lutraconsulting dot co dot uk
  ***************************************************************************
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
@@ -13,52 +13,44 @@
  *                                                                         *
  ***************************************************************************/
 
-#ifndef QGS3DMAPTOOLPAINTBRUSH_H
-#define QGS3DMAPTOOLPAINTBRUSH_H
-#include "qgs3dmaptool.h"
-#include "qgspoint.h"
+#ifndef QGS3DMAPTOOLPOLYGON_H
+#define QGS3DMAPTOOLPOLYGON_H
+
 #include "qgs3dmaptoolpointcloudchangeattribute.h"
 
+#include <QPointF>
+
 class QgsRubberBand3D;
-class Qgs3DMapToolPaintBrush : public Qgs3DMapToolPointCloudChangeAttribute
+class QgsPointXY;
+
+class Qgs3DMapToolPolygon : public Qgs3DMapToolPointCloudChangeAttribute
 {
     Q_OBJECT
 
   public:
-    explicit Qgs3DMapToolPaintBrush( Qgs3DMapCanvas *canvas );
-    ~Qgs3DMapToolPaintBrush() override;
+    Qgs3DMapToolPolygon( Qgs3DMapCanvas *canvas );
+    ~Qgs3DMapToolPolygon() override;
 
     void activate() override;
-
     void deactivate() override;
 
-    QCursor cursor() const override;
-
     bool allowsCameraControls() const override { return false; }
-
-    void restart() override;
 
   private slots:
     void mousePressEvent( QMouseEvent *event ) override;
     void mouseReleaseEvent( QMouseEvent *event ) override;
     void mouseMoveEvent( QMouseEvent *event ) override;
-    void mouseWheelEvent( QWheelEvent *event ) override;
     void keyPressEvent( QKeyEvent *event ) override;
 
   private:
     void run() override;
-    void generateHighlightArea();
+    void restart() override;
 
-    std::unique_ptr<QgsRubberBand3D> mSelectionRubberBand;
-    std::unique_ptr<QgsRubberBand3D> mHighlighterRubberBand;
-    QVector<QgsPointXY> mDragPositions = QVector<QgsPointXY>();
-    //! Check if mouse was moved between mousePress and mouseRelease
-    bool mIsClicked = false;
-    //! Check if the tool is selected
-    bool mIsActive = false;
-    //! Check if the tool or movement is active
+    QVector<QgsPointXY> mScreenPoints;
+    QgsRubberBand3D *mPolygonRubberBand = nullptr;
+    QPoint mClickPoint;
     bool mIsMoving = false;
 };
 
 
-#endif //QGS3DMAPTOOLPAINTBRUSH_H
+#endif //QGS3DMAPTOOLPOLYGON_H
